refactor(api): split context.rs into domain sub-modules
ServerContext implementation split into 7 focused files:
- mod.rs: struct, new(), impl Context
- player.rs: impl PlayerContext
- chat.rs: impl ChatContext
- world.rs: impl WorldContext
- entity.rs: impl EntityContext
- container.rs: impl ContainerContext
- response.rs: Response enum + ResponseQueue
- tests.rs: unit tests

Removes teleport_counter field — uses the static directly.
